GH-20: Change the file structure by splitting the project into two parts
