+++ rollup.config.ts
@@ -6,10 +6,14 @@ import {readFileSync} from "fs";
 
 const pkg = JSON.parse(readFileSync("./package.json") as unknown as string);
 
-const input = "src/index.tsx";
-const cjsOutput = {file: pkg.main, format: "cjs", exports: "auto"};
-const esmOutput = {file: pkg.module, format: "es"};
-const dtsOutput = {file: pkg.types, format: "es"};
+const input4 = "src/Input4/index.tsx";
+const input5 = "src/Input5/index.tsx";
+const cjsInput4 = {file: "Input4/index.cjs.js", format: "cjs", exports: "auto"};
+const esmInput4 = {file: "Input4/index.esm.js", format: "es"};
+const dtsInput4 = {file: "Input4/index.d.ts", format: "es"};
+const cjsInput5 = {file: "Input5/index.cjs.js", format: "cjs", exports: "auto"};
+const esmInput5 = {file: "Input5/index.esm.js", format: "es"};
+const dtsInput5 = {file: "Input5/index.d.ts", format: "es"};
 
 const jsonPlugin = json();
 const cssPlugin = postcss();
@@ -23,7 +27,10 @@ const external = [
 ];
 
 export default [
-	{input, output: cjsOutput, plugins: [tsPlugin, jsonPlugin, cssPlugin], external},
-	{input, output: esmOutput, plugins: [tsPlugin, jsonPlugin, cssPlugin], external},
-	{input, output: dtsOutput, plugins: [dts()], external: [/\.css$/]},
+	{input: input4, output: cjsInput4, plugins: [tsPlugin, jsonPlugin, cssPlugin], external},
+	{input: input4, output: esmInput4, plugins: [tsPlugin, jsonPlugin, cssPlugin], external},
+	{input: input4, output: dtsInput4, plugins: [dts()], external},
+	{input: input5, output: cjsInput5, plugins: [tsPlugin, jsonPlugin, cssPlugin], external},
+	{input: input5, output: esmInput5, plugins: [tsPlugin, jsonPlugin, cssPlugin], external},
+	{input: input5, output: dtsInput5, plugins: [dts()], external},
 ];
